Fix RefCne to check prior TurnOn/Off request status
Inlcudes fixes to remove infinite loop of Turn On requests issued by RefCne
to ConnectivityService.
Fix to correct the debug message printed to logcat
Changed logcat macro to print all RefCne messages to ERROR to avoid recompiling
of an engineering build to do the same for test purposes.
Added return to after the check for invalid net preference
Fixed typo in the debug msg printed to logcat in SetPreferredNetwork

Change-Id: Id2d96176c2dfff69a2cd2937ad3960c1a7eb5061
CRs-fixed: 229108,227945